feat-28 add chart container and refactor relevant files.
