+++ public/index.html
@@ -11,8 +11,8 @@
     integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA=="
     crossorigin=""></script>
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="stylesheet" href="bower_components/aos/dist/aos.css" />
-  <script src="bower_components/aos/dist/aos.js"></script>
+  <!-- <link rel="stylesheet" href="bower_components/aos/dist/aos.css" /> -->
+  <!-- <script src="bower_components/aos/dist/aos.js"></script> -->
   <meta name="description" content="Budget Application" />
   <title>Budget Application</title>
 </head>
