+++ src/components/organisms/Charts/Charts.tsx
@@ -1,19 +1,27 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import React, { FC } from 'react';
 import Chart from 'react-apexcharts';
-import { lineChartOptions, lineChartSeries, barChartSeries, barChartOptions } from '../../../constants/charts';
+
 import * as S from './style';
+
+interface IChart {
+  lineChartOptions?: any;
+  lineChartSeries?: any;
+  barChartOptions?: any;
+  barChartSeries?: any;
+}
+
 interface IChartsProps {
-  [property: string]: FC;
+  [property: string]: FC<IChart>;
 }
 
 const Charts: IChartsProps = {
-  Income: () => (
+  Income: ({ lineChartOptions, lineChartSeries }) => (
     <S.LineChartWrapper>
       <Chart options={lineChartOptions} series={lineChartSeries} type="line" />
     </S.LineChartWrapper>
   ),
-  Activity: () => (
+  Activity: ({ barChartOptions, barChartSeries }) => (
     <S.BarChartWrapper>
       <Chart options={barChartOptions} series={barChartSeries} type="bar" height={350} />
     </S.BarChartWrapper>
