+++ src/containers/ChartContainer/index.ts
@@ -1,0 +1,2 @@
+import ChartContainer from './ChartContainer';
+export default ChartContainer;
