+++ src/containers/index.ts
@@ -1,2 +1,3 @@
 export { default as DashboardContainer } from './DashboardContainer';
 export { default as TransactionsContainer } from './TransactionsContainer';
+export { default as ChartContainer } from './ChartContainer';
